Reverting earlier changes to make a direction.rei
Justification: the direction type is primarily used for pattern matching, and needs to be exposed.

See: http://dev.realworldocaml.org/files-modules-and-programs.html

Despite these benefits, there is a trade-off here. In particular, exposing types concretely makes it possible to use pattern-matching with those types, which as we saw in Lists And Patterns is a powerful and important tool. You should generally only expose the concrete implementation of your types when there's significant value in the ability to pattern match, and when the invariants that you care about are already enforced by the data type itself.